// src/app/core/class/models/menu.permission.ts
export class MenuPermission {
    IsView: any;
    IsCreate: any;
    IsUpdate: any;
    IsDelete:  any;
    /**
     * name
     */
    // public constructor(isView: any, isCreate:any, isUpdate: any, isDelete: any) {
    //   this.IsView = isView;
    //   this.IsCreate = isCreate;
    //   this.IsUpdate = isUpdate;
    //   this.IsDelete = isDelete;
    // }
  }